# level 5 assignment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/karyawan": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Karyawan"
                ],
                "summary": "Put Karyawan Data",
                "parameters": [
                    {
                        "description": "Karyawan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateKaryawanRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new Karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Karyawan"
                ],
                "summary": "Post Karyawan Data",
                "parameters": [
                    {
                        "description": "Karyawan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SaveKaryawanRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/karyawan/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Karyawan"
                ],
                "summary": "Get All Karyawan Data",
                "responses": {}
            }
        },
        "/karyawan/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Karyawan"
                ],
                "summary": "Get Karyawan Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Karyawan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Karyawan"
                ],
                "summary": "Delete Karyawan Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Karyawan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/karyawanTraining": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update KaryawanTraining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KaryawanTraining"
                ],
                "summary": "Put KaryawanTraining Data",
                "parameters": [
                    {
                        "description": "KaryawanTraining",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateKaryawanTrainingRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new KaryawanTraining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KaryawanTraining"
                ],
                "summary": "Post KaryawanTraining Data",
                "parameters": [
                    {
                        "description": "KaryawanTraining",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SaveKaryawanTrainingRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/karyawanTraining/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All KaryawanTraining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KaryawanTraining"
                ],
                "summary": "Get All KaryawanTraining Data",
                "responses": {}
            }
        },
        "/karyawanTraining/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get KaryawanTraining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KaryawanTraining"
                ],
                "summary": "Get KaryawanTraining Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete KaryawanTraining",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KaryawanTraining"
                ],
                "summary": "Delete KaryawanTraining Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "KaryawanTraining ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/user-login/login": {
            "post": {
                "description": "Login Request Token for Authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Request Token user",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "dtos.KaryawanRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dtos.SaveDetailKaryawan": {
            "type": "object",
            "properties": {
                "nik": {
                    "type": "string"
                },
                "npwp": {
                    "type": "string"
                }
            }
        },
        "dtos.SaveKaryawanRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "detailKaryawan": {
                    "$ref": "#/definitions/dtos.SaveDetailKaryawan"
                },
                "dob": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.SaveKaryawanTrainingRequest": {
            "type": "object",
            "properties": {
                "karyawan": {
                    "$ref": "#/definitions/dtos.KaryawanRequest"
                },
                "tanggal": {
                    "type": "string"
                },
                "training": {
                    "$ref": "#/definitions/dtos.TrainingRequest"
                }
            }
        },
...
        "dtos.TrainingRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateDetailKaryawan": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "npwp": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateKaryawanRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "detailKaryawan": {
                    "$ref": "#/definitions/dtos.UpdateDetailKaryawan"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateKaryawanTrainingRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "karyawan": {
                    "$ref": "#/definitions/dtos.KaryawanRequest"
                },
                "tanggal": {
                    "type": "string"
                },
                "training": {
                    "$ref": "#/definitions/dtos.TrainingRequest"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}